% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_formula.R
\name{create_formula}
\alias{create_formula}
\title{Create formula from variable names}
\usage{
create_formula(ind_var, dep_var, intercept = TRUE)
}
\arguments{
\item{ind_var}{string of independent(y) variable name}

\item{dep_var}{string or list of strings of dependent(X1, X2, ...) variable names}

\item{intercept}{boolean. When FALSE intercept is 0}
}
\value{
formula object for use in lm() and similar functions
}
\description{
create a formula type object from lists of variable names
for use in the lm() function. Called by regress_group_data() and may not be useful for end users.
}
\examples{
create_formula(ind_var = "happiness", dep_var = c("cats","vacation_time"))
> "happiness ~ cats + vacation_time"
}
