% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_spooky_time.R
\name{is_spooky_time}
\alias{is_spooky_time}
\title{Is it spooky time?}
\usage{
is_spooky_time(today = Sys.Date())
}
\arguments{
\item{today}{date or a string which can be coerced into a date. by default uses current date.
Date format can be mm/dd/yyyy or yyyy/mm/dd. and use '-' or '/" as separators}
}
\value{
boolean. TRUE when it is a spooky time of year
}
\description{
Is it spooky time?
}
\examples{
is_spooky_time()
is_spooky_time(as.Date("2020-10-31"))
}
