% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_mp3.R
\name{play_mp3}
\alias{play_mp3}
\title{Play MP3 from assets folder}
\usage{
play_mp3(song)
}
\arguments{
\item{song}{string. name of .mp3 file in assets folder.}
}
\description{
this is used internally in is_spooky_time() and not for end users.
}
\examples{
play_mp3("spookysong")
}
