% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_group_data.R
\name{regress_group_data}
\alias{regress_group_data}
\title{Regress group data}
\usage{
regress_group_data(
  data,
  group_id = NULL,
  obs,
  cov,
  include_intercept = TRUE,
  ...
)
}
\arguments{
\item{data}{a dataframe}

\item{group_id}{a string or list of strings representing the group id column}

\item{obs}{a string representing the column name of the observed variable}

\item{cov}{a string or list of strings representing covariates}

\item{include_intercept}{boolean. whether to include intercept, TRUE by default}

\item{...}{other arguments}
}
\value{
dataframe containing parameters for linear regression
}
\description{
Regress group data
}
\examples{
regress_group_data(mtcars, "gear", "mpg", "disp")
}
