% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_group_data.R
\name{summarise_group_data}
\alias{summarise_group_data}
\title{Summarise Group Data}
\usage{
summarise_group_data(data, group_id = NULL, obs, fun, ...)
}
\arguments{
\item{data}{a data frame}

\item{group_id}{a string or vector of strings of column names to group by}

\item{obs}{a string or vector of strings of column names to summarise}

\item{fun}{a function to summarise with or a list of functions}

\item{...}{any additional arguments for the function}
}
\value{
a dataframe containing the summary of the obs variable
}
\description{
Summarise Group Data
}
\examples{
summarise_group_data(mtcars, "gear", "mpg", mean)
summarise_group_data(mtcars, "gear", c("disp","mpg"), mean)
}
