% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_group_data.R
\name{summarize_group_data}
\alias{summarize_group_data}
\title{Summarize group data}
\usage{
summarize_group_data(data, group_id = NULL, obs, fun, ...)
}
\description{
Identical to summarise_group_data. View documentation for that function.
}
