% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_g_d_inputs.R
\name{valid_g_d_inputs}
\alias{valid_g_d_inputs}
\title{Valid Group Data inputs}
\usage{
valid_g_d_inputs(group_id, other_vars)
}
\arguments{
\item{group_id}{character, list of characters, or null}

\item{other_vars}{character or list of characters}
}
\value{
boolean. if TRUE, inputs are valid
}
\description{
Used in regress_group_data and summarise_group_data
to determine whether function inputs have the below structure.
}
\examples{
Not for end user.
}
